Changes made to remove the warning
Though it was not exactly clear why it is throwing the error, considering the fact that, I have checked the code again for any argument or variable mismatch to the best of my knowledge. The only conclusion, I can reach was some ussue with the XAMP setup and somebug in PHP 5. 

Just rewrote the getMeaning.php code, using Object Oriented approch of PHP 5. This resoved the issue. Please do revert back, if there are anyother warnings.